% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccs_tree.R
\name{ccs_tree}
\alias{ccs_tree}
\title{\code{ccs_tree} Produces tree of Clinical Classification Software (CCS) codes
based on root node.}
\usage{
ccs_tree(group = NULL)
}
\arguments{
\item{group}{character string specifying that only codes beginning with group 
will be included in tree The default is NULL (returns full tree of CCS codes)}
}
\value{
A list containing the following elements:
tr = directed igraph object; the tree of outcomes.
ccs_icd_mapping = data frame specifying the ICD9 codes corresponding to each 
CCS code. Useful for converting from ICD9 to CCS.
}
\description{
All the details go here!
}
\section{Details}{
NA
}

\concept{tree functions}
