% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_by_group.R
\name{ml_by_group}
\alias{ml_by_group}
\title{Here's a brief description.
  \code{spike_and_slab_normal} performs group variable selection via a spike
  and slab prior. The posterior is approximated via variational inference.
  This function returns coefficient estimates and 95% credible intervals.}
\usage{
ml_by_group(X, W = NULL, y, outcomes, outcome_groups, ci_level, family)
}
\arguments{
\item{X}{An n x K matrix of exposure data, where K is the dimension of the exposure.
Grouping of the outcomes will be based on their relationships with the variables in X.}

\item{W}{Matrix of covariates of dimension n x m.
Coefficients for these variables do not affect grouping of the outcomes.}

\item{y}{Vector of length n containing outcomes data.
If family = "bernoulli", y must be an integer vector where 1 = success, 0 = failure.
If family = "gaussian", y must be a numeric vector containing continuous data.}

\item{outcomes}{Character vector specifying which outcomes the elements of y 
(rows of X and W) correspond to.}

\item{outcome_groups}{A list of length G, where each element is a character vector
indiciating which outcomes belong to that group.}

\item{ci_level}{A number between 0 and 1 giving the desired credible interval.
For example, ci_level = 0.95 (the default) returns a 95% credible interval.}

\item{family}{A string specifying the distribution of the outcomes: 
either "bernoulli" (for classification) or "gaussian" (for regression)}
}
\value{
A list containing the following elements:
1. estimated coefficients and credible intervals; 
2. outputs from variational inference algorithm
}
\description{
All the details go here!
}
\section{Model Description}{

  Describe group spike and slab prior and all parameters here.
}

\seealso{
Other spike and slab functions: 
\code{\link{compute_betas}()},
\code{\link{compute_thetas}()},
\code{\link{spike_and_slab_logistic_moretrees}()},
\code{\link{spike_and_slab_logistic}()},
\code{\link{spike_and_slab_normal_moretrees}()},
\code{\link{spike_and_slab_normal}()},
\code{\link{spike_and_slab}()}
}
\concept{spike and slab functions}
