% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moretrees_compute_thetas.R
\name{moretrees_compute_thetas}
\alias{moretrees_compute_thetas}
\title{Here's a brief description.
\code{moretrees_compute_thetas} Computes MOReTreeS estimates and credible
  intervals for covariate effects from model output}
\usage{
moretrees_compute_thetas(
  mod,
  ci_level,
  W_method = "shared",
  method = "tree",
  m,
  A_leaves = NULL
)
}
\arguments{
\item{mod}{List containing outputs from spike and slab VI algorithm}

\item{ci_level}{A number between 0 and 1 giving the desired credible interval.
For example, ci_level = 0.95 (the default) returns a 95\% credible interval.}

\item{W_method}{A string, either "shared" or "individual"}

\item{m}{Integer number of variables in covariate matrix W}

\item{A_leaves}{pL x p sparse ancestor Matrix where rows correspond to leaves
of tree (outcomes) and columns correspond to nodes on tree. Results in mod
must have same ordering as columns of A_leaves.}
}
\value{
A matrix containing the estimates and confidence intervals.
}
\description{
All the details go here!
}
\section{Details}{
NA
}

\seealso{
Other MOReTreeS functions: 
\code{\link{moretrees_compute_betas}()},
\code{\link{moretrees_design_matrix}()},
\code{\link{moretrees_design_tree}()},
\code{\link{moretrees_init_W_logistic}()},
\code{\link{moretrees_init_logistic}()},
\code{\link{moretrees_init_rand}()},
\code{\link{moretrees}()}
}
\concept{MOReTreeS functions}
