% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moretrees_init_logistic.R
\name{moretrees_init_logistic}
\alias{moretrees_init_logistic}
\title{Here's a brief description.
  \code{moretrees_init_rand} Randomly generates starting values for moretrees
  models. Not recommended if the model is converging slowly!}
\usage{
moretrees_init_logistic(
  X,
  W,
  y,
  A,
  outcomes_units,
  outcomes_nodes,
  ancestors,
  xxT,
  wwT,
  update_hyper,
  hyper_fixed
)
}
\arguments{
\item{xxT}{Computed from exposure matrix X}

\item{wwT}{Computed from covariate matrix W}

\item{update_hyper}{Update hyperparameters? Default = TRUE.}

\item{hyper_fixed}{Fixed values of hyperparameters to use if update_hyper = FALSE.
If family = "bernoulli", this should be a list including the following elements:
tau (prior variance for sparse node coefficients)
rho (prior node selection probability for sparse node coefficients)
omega (prior variance for non-sparse node coefficients)
If family = "gaussian", in addition to the above, the list should also include:
sigma2 (variance of residuals)}

\item{dsgn}{Design list generated by moretrees_design_tree()}

\item{hyper_random_init}{If update_hyper = TRUE, this is a list containing the 
maximum values of the hyperparameters. Each hyperparameter will be initialised
uniformly at random between 0 and the maximum values given by the list elements
below. If multiple random restarts are being used, it is recommended
to use a large range for these initial values so that the parameter space
can be more effectively explored. The list contains the following elements:
tau_max (maxmimum of prior sparse node variance)
omega_max (maximum of prior non-sparse node variance)
sigma2_max (maximum of residual error variance--- for gaussian data only)}

\item{vi_random_init}{A list with parameters that determine the distributions from
which the initial VI parameters will be randomly chosen. All parameters will be randomly
selected from independent normal distributions with the standard deviations given by
the list elements below. If multiple random restarts are being used, it is recommended
to use large standard deviations for these initial values so that the parameter space
can be more effectively explored. The list contains the following elements:
mu_sd (standard deviation for posterior means of sparse node coefficients)
delta_sd (standard deviation for posterior means of non-sparse node coefficients)
xi_sd (standard deviation for auxilliary parameters xi--- for bernoulli data only)}
}
\value{
A list containing starting values
}
\description{
Here's a brief description.
  \code{moretrees_init_rand} Randomly generates starting values for moretrees
  models. Not recommended if the model is converging slowly!
}
\section{Model Description}{

Describe MOReTreeS model and all parameters here.
}

\seealso{
Other MOReTreeS functions: 
\code{\link{moretrees_compute_betas}()},
\code{\link{moretrees_compute_thetas}()},
\code{\link{moretrees_design_matrix}()},
\code{\link{moretrees_design_tree}()},
\code{\link{moretrees_init_W_logistic}()},
\code{\link{moretrees_init_rand}()},
\code{\link{moretrees}()}
}
\concept{MOReTreeS functions}
