% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spike_and_slab_logistic.R
\name{spike_and_slab_logistic}
\alias{spike_and_slab_logistic}
\title{Group spike and slab variable selection with Gaussian outcome}
\usage{
spike_and_slab_logistic(
  dsgn,
  initial_values,
  tol,
  max_iter,
  update_hyper,
  update_hyper_freq,
  hyper_fixed,
  print_freq,
  hyper_random_init,
  vi_random_init
)
}
\arguments{
\item{tol}{Convergence tolerance for ELBO.}

\item{update_hyper}{Update hyperparameters? Default = TRUE.}

\item{update_hyper_freq}{How frequently to update hyperparameters. Default = every 10 iterations.}

\item{y}{Integer vector of length n containing outcomes; 1 = success, 0 = failure.}

\item{X}{matrix of dimension n x sum(K), where n is the number of units, and
K[g] is the number of variables in group g.}

\item{groups}{A list of length G (number of groups), where groups[[g]] is an integer
vector specifying the columns of X that belong to group g.}

\item{W}{matrix of non-sparse regression coefficients of dimension n x m}

\item{maxiter}{Maximum number of iterations of the VI algorithm.}
}
\value{
A list of variational parameters.
}
\description{
Here's a brief description.
  \code{spike_and_slab_logistic} performs group variable selection via a spike
  and slab prior for binary data.
  The posterior is approximated via variational inference.
  This function returns the parameters of the variational approximation.
}
\details{
All the details go here!
}
\section{Model Description}{

  Describe group spike and slab prior and all parameters here.
}

\seealso{
Other spike and slab functions: 
\code{\link{compute_betas}()},
\code{\link{compute_thetas}()},
\code{\link{ml_by_group}()},
\code{\link{spike_and_slab_logistic_moretrees}()},
\code{\link{spike_and_slab_normal_moretrees}()},
\code{\link{spike_and_slab_normal}()},
\code{\link{spike_and_slab}()}
}
\concept{spike and slab functions}
