% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spike_and_slab_normal.R
\name{spike_and_slab_normal}
\alias{spike_and_slab_normal}
\title{Group spike and slab variable selection with Gaussian outcome}
\usage{
spike_and_slab_normal(
  dsgn,
  initial_values,
  tol,
  max_iter,
  update_hyper,
  update_hyper_freq,
  hyper_fixed,
  print_freq,
  hyper_random_init,
  vi_random_init
)
}
\arguments{
\item{tol}{Convergence tolerance for ELBO.}

\item{max_iter}{Maximum number of iterations of the VI algorithm.}

\item{update_hyper}{Update hyperparameters? Default = TRUE.}

\item{update_hyper_freq}{How frequently to update hyperparameters. 
Default = every 50 iterations.}

\item{print_freq}{How often to print out iteration number.}

\item{y}{Numeric vector of length n of outcome data}

\item{X}{Matrix of dimension n x sum(K), where n is the number of units, and
K[g] is the number of variables in group g.}

\item{groups}{A list of length G (number of groups), where groups[[g]] is an integer
vector specifying the columns of X that belong to group g.}

\item{W}{Matrix of data with non-sparse regression coefficients of dimension n x m}

\item{nrestarts}{Number of random re-starts of the VI algorithm. The result that 
gives the highest ELBO will be returned. It is recommended to choose nrestarts > 1.}
}
\value{
A list of variational parameters.
}
\description{
Here's a brief description.
  \code{spike_and_slab_normal} performs group variable selection via a spike
  and slab prior for continuous, normally distributed data.
  The posterior is approximated via variational inference.
  This function returns the parameters of the variational approximation.
}
\details{
All the details go here!
}
\section{Model Description}{

  Describe group spike and slab prior and all parameters here.
}

\seealso{
Other spike and slab functions: 
\code{\link{compute_betas}()},
\code{\link{compute_thetas}()},
\code{\link{ml_by_group}()},
\code{\link{spike_and_slab_logistic_moretrees}()},
\code{\link{spike_and_slab_logistic}()},
\code{\link{spike_and_slab_normal_moretrees}()},
\code{\link{spike_and_slab}()}
}
\concept{spike and slab functions}
