% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_hyperparams_logistic_moretrees.R
\name{update_hyperparams_logistic_moretrees}
\alias{update_hyperparams_logistic_moretrees}
\title{\code{update_hyperparams_logistic_moretrees} Performs hyperparameter updates and computes 
  current value of ELBO in VI algorithm for bernoulli outcomes.}
\usage{
update_hyperparams_logistic_moretrees(
  X,
  W,
  y,
  outcomes_units,
  ancestors,
  n,
  K,
  p,
  m,
  prob,
  mu,
  Sigma,
  Sigma_det,
  tau_t,
  delta,
  Omega,
  Omega_det,
  eta,
  g_eta,
  a_rho,
  b_rho,
  omega,
  tau,
  model = "ss",
  update_hyper = T
)
}
\description{
\code{update_hyperparams_logistic_moretrees} Performs hyperparameter updates and computes 
  current value of ELBO in VI algorithm for bernoulli outcomes.
}
