% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_hyperparams_normal.R
\name{update_hyperparams_normal}
\alias{update_hyperparams_normal}
\title{\code{update_hyperparams_normal} Performs hyperparameter updates and computes 
  current value of ELBO in VI algorithm for gaussian outcomes}
\usage{
update_hyperparams_normal(
  X,
  groups,
  XtX,
  W,
  WtW,
  y,
  n,
  K,
  G,
  m,
  prob,
  mu,
  Sigma,
  Sigma_det,
  tau_t,
  delta,
  Omega,
  Omega_det,
  omega,
  sigma2,
  tau,
  rho,
  update_hyper = T
)
}
\description{
\code{update_hyperparams_normal} Performs hyperparameter updates and computes 
  current value of ELBO in VI algorithm for gaussian outcomes
}
